Make firmware generic across keyboards
This commit moves all keyboard specifics into a the machines/
folder. This includes sizing of arrays, pin definitions, subsystem
parameters, USB vid/pid, number of layers, etc

Several of the functions in keyboard.c and other subsystems are now
weak functions, allowing the specific keyboard to supply
an implementation.

Subsystem arrays like combos and macros are marked as extern in the main
firmware and must be supplied by the keyboard.
