+++ src/combo.h
@@ -9,12 +9,6 @@
 #include "pico/types.h"
 #include "keyboard.h"
 
-// defines
-#define NUM_COMBO_SLOTS             (16)
-#define MAX_KEYS_PER_COMBO          (4)
-#define COMBO_DELAY_MS              (50)
-#define COMBO_CANCEL_SUPPRESS_MS    (150)
-
 // typedefs
 typedef struct rowcol_t {
     uint8_t row;
