+++ src/leds.h
@@ -9,9 +9,6 @@
 #include "pico/types.h"
 #include "keyboard.h"
 
-// defines
-#define NUM_LEDS  (2)
-
 // typedefs
 typedef struct leds_state_t {
     uint8_t leds[NUM_LEDS][3];
@@ -32,5 +29,3 @@ void leds_init(void);
 void leds_brightness_up(void);
 void leds_brightness_down(void);
 void leds_toggle_led_enabled(uint led_index);
-
-void leds_write_test_led(bool on);
