+++ src/machines/machine.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#ifdef KEYBOARD_HEX2A
+#include "hex2a/configuration.h"
+#endif
